requirements.txt
langchain == 0.3.27
langchain-community == 0.3.27
langchain-chroma == 0.2.5
langchain-core == 0.3.72
pypdf == 5.0.0
langchain-openai == 0.3.28
chainlit == 2.6.3
python-dotenv == 1.0.1
sentence-transformers == 3.0.1
langgraph == 0.6.0
numpy == 1.26.4
pydantic == 2.11.7
